fix(ci): unbreak all four gates broken since v1.11.0 release
CI was failing on main and on every Dependabot PR because of four
unrelated regressions that landed (or got exposed) by the public
release commit.

* mypy: workflow targeted core.py but the file had been split into a
  core/ package on 2026-04-28; switch the arg list to the package.
* ruff: 1112 violations, 815 of them RUF001/002/003 false-positives
  on Turkish characters in a TR-language codebase. Globally ignore
  those three; per-file-ignore F401 in __init__.py (re-exports),
  E402 in app.py (sys.path bootstrap before imports), and B904/E702
  /RUF043 in tests/. Fix the remaining genuine issues by hand
  (raise … from None for HTTPException translations in routers/batch,
  rename ambiguous l→lf, drop semicolons, swap completed counter for
  enumerate, add context manager to one open()). Run ruff format on
  the tree.
* test_pdf_editor: NBSP (U+00A0) leaks out of pymupdf get_text() on
  certain font combinations; normalise the substring assertion so
  Türkçe/Yeni Metin checks pass everywhere. Switch the noto-sans
  catalog test to the same on-disk guard the rest of the file uses.
* test_security: safe_filename collapsed only \, /, etc. — but on
  POSIX the backslash in '..\foo.pdf' is literal, so Path().stem
  passed '..\foo' through and the substitution left '.._foo' with
  '..' intact. Defense-in-depth: collapse any surviving '..' runs.

Two latent type errors surfaced once mypy could read the package
again — fixed in core/editor.py (bool() coerce) and
core/clamav_update.py (explicit dict annotation).

Verified locally: ruff check & format clean, mypy success on 42
files, packaging gate ok, pytest 394 passed / 1 skipped (Windows
admin-only symlink test).

+++ core/distribution.py
@@ -41,7 +41,7 @@ def distribute_custom(
     total = sum(ratios)
     if total <= 0:
         return out
-    counts = [int(round(len(records) * r / total)) for r in ratios]
+    counts = [round(len(records) * r / total) for r in ratios]
     diff = len(records) - sum(counts)
     if diff != 0:
         counts[0] += diff
